+++ app/Http/Controllers/Teacher/TaskController.php
@@ -19,6 +19,7 @@ use App\Models\TaskAssignee;
 use App\Models\User;
 use App\Models\Users\StudentUser;
 use App\Models\Users\TeacherUser;
+use App\Services\TaskReaderService;
 use App\Traits\Common;
 use App\Traits\LogActivity;
 use App\Traits\TodolistProcess;
@@ -41,6 +42,8 @@ class TaskController extends Controller
     use Common;
     use LogActivity;
     use TodolistProcess;
+
+    public function __construct(protected TaskReaderService $taskReader) {}
 
     /**
      * Display task list grouped by task flag (API).
@@ -279,7 +282,11 @@ class TaskController extends Controller
      */
     public function show($id)
     {
-        $task = Task::where('id', $id)->firstOrFail();
+        $task = $this->taskReader->find($id, Auth::user()->school_id);
+
+        if (! $task) {
+            abort(404);
+        }
 
         $task_assignees = TaskAssignee::with(['user', 'standardLink'])
             ->where('task_id', $id)
@@ -390,8 +397,12 @@ class TaskController extends Controller
      */
     public function editList(Request $request, $id)
     {
-        //
-        $task = Task::where('id', $id)->first();
+        $task = $this->taskReader->find($id, Auth::user()->school_id);
+
+        if (! $task) {
+            abort(404);
+        }
+
         $task_assignees = TaskAssignee::where('task_id', $id)->get();
 
         foreach ($task_assignees as $key => $task_assignee) {
@@ -443,8 +454,11 @@ class TaskController extends Controller
      */
     public function edit($id)
     {
-        //
-        $task = Task::where('id', $id)->first();
+        $task = $this->taskReader->find($id, Auth::user()->school_id);
+
+        if (! $task) {
+            abort(404);
+        }
 
         return view('/teacher/todolist/edit', ['task' => $task]);
     }
@@ -457,12 +471,17 @@ class TaskController extends Controller
      */
     public function update(TaskRequest $request, $id)
     {
+        $school_id = Auth::user()->school_id;
+
+        if (! $this->taskReader->find($id, $school_id)) {
+            abort(404);
+        }
+
         try {
-            $school_id = Auth::user()->school_id;
             $academic_year = SiteHelper::getAcademicYear($school_id);
             $auth_id = Auth::id();
 
-            $task = $this->editTaskAssignee($request, $auth_id, $id);
+            $task = $this->editTaskAssignee($request, $auth_id, $id, $school_id);
 
             $message = trans('messages.update_success_msg', ['module' => 'Task']);
 
@@ -491,13 +510,18 @@ class TaskController extends Controller
      */
     public function snooze(Request $request, $id)
     {
+        $school_id = Auth::user()->school_id;
+        $task = $this->taskReader->find($id, $school_id);
+
+        if (! $task) {
+            abort(404);
+        }
+
         try {
-            $school_id = Auth::user()->school_id;
             $academic_year = SiteHelper::getAcademicYear($school_id);
             $auth_id = Auth::id();
-            $task = Task::where('id', $id)->first();
             if ($task->snooze == 0) {
-                $task = $this->snoozeTask($request, $auth_id, $id);
+                $task = $this->snoozeTask($request, $auth_id, $id, $school_id);
 
                 $mins = env('SNOOZE_TIME') / 60;
                 $message = trans('messages.task_snooze_msg', ['mins' => $mins]);
@@ -530,9 +554,13 @@ class TaskController extends Controller
      */
     public function destroy($id)
     {
-        try {
-            $task = Task::where('id', $id)->first();
+        $task = $this->taskReader->find($id, Auth::user()->school_id);
 
+        if (! $task) {
+            abort(404);
+        }
+
+        try {
             $task->delete();
 
             $message = trans('messages.delete_success_msg', ['module' => 'Task']);
